Whole-package readability pass: fix doc rot, document pad fields, loader cleanups
Three parallel reviewers (reuse / readability / efficiency+altitude)
over the ENTIRE package including the never-reviewed python module,
tools, tests, ply_loader.rs and nerf_camera.rs. Applied:

Readability (the priority - this code will be human-reviewed):
- gpu_types.rs: document every _pad field. ProjectUniformBuffer._pad is
  NOT free padding (bit-packed has-SH flag + OPACITY_SCALE bits, read
  as project_uniforms.pad.x/.y in WGSL) - previously an invisible
  deletion trap; the others are genuine WGSL alignment padding and now
  say so.
- render_cli.rs: header claimed a '7-stage' pipeline; now states it
  shares stages 1-8 of the 9-stage pipeline with readback in place of
  the viewer-only composite.
- docs/architecture.md: remove deleted covariance.rs from the module
  box / dependency graph / file map, renumber the shader file map to
  the 9-stage scheme, fix layout/pipeline counts (12/13 after the dead
  sort_scan removal), update sh.rs description.
- README.md: Cargo.toml line no longer claims published rerun 0.33.0;
  new note documents the local-checkout path deps, the mirrored egui
  patches, and the PKG_CONFIG_PATH requirement for bare cargo builds.
- docs/presentation.html: prominent historical-snapshot banner (the
  deck narrates the pre-port CPU-pre-pass architecture; full refresh
  deferred as its own task).

Efficiency (never-reviewed files):
- ply_loader.rs: hoist the f_rest_* field-name Strings out of the
  per-vertex loop (was 3 allocations per coefficient per vertex).
- gaussians3d.py: drop the redundant astype copy after _sigmoid
  (already float32).

Reviewers' no-flag verdicts kept on record: no duplication findings
package-wide; accepted-architecture items not re-litigated.

Gates: spot-check renders + calibration + 310k bit-identical (loader
change sits on the load path), relog check PASS, clippy -D warnings
both feature sets, 11 Rust + 13 python tests, ruff clean.

+++ packages/gsplat-rust-renderer/gsplat_rust_renderer/gaussians3d.py
@@ -281,9 +281,7 @@ class Gaussians3D(rr.AsComponents):
         quaternions_xyzw: Float32[np.ndarray, "n 4"] = np.stack(
             [vertex["rot_1"], vertex["rot_2"], vertex["rot_3"], vertex["rot_0"]], axis=1
         ).astype(np.float32)
-        opacities: Float32[np.ndarray, "n"] = _sigmoid(
-            np.asarray(vertex["opacity"], dtype=np.float32)
-        ).astype(np.float32)
+        opacities: Float32[np.ndarray, "n"] = _sigmoid(np.asarray(vertex["opacity"], dtype=np.float32))
 
         dc_coefficients: Float32[np.ndarray, "n 3"] | None = None
         if {"f_dc_0", "f_dc_1", "f_dc_2"} <= names:
